+++ include/brain/Layer.hh
@@ -31,13 +31,13 @@ struct LayerParam
     {
     }
 
-    unsigned size; //number of neurons
+    size_t size; //number of neurons
     double maxNorm;
     Distrib distrib;
     double mean_boundary; //mean (if normal), boundary (if uniform)
     double deviation; //deviation (if normal) or useless (if uniform)
     bool useOutput; // calculate boundary/deviation by taking output number into account
-    unsigned k; //number of weight set for each neuron (for maxout)
+    size_t k; //number of weight set for each neuron (for maxout)
 };
 
 
@@ -58,8 +58,8 @@ public:
     virtual Matrix processToLearn(Matrix const& inputs, double dropout, double dropconnect, std::bernoulli_distribution& dropoutDist, std::bernoulli_distribution& dropconnectDist, std::mt19937& dropGen, ThreadPool& t) = 0;
     virtual void computeGradients(Matrix const& inputGradients, ThreadPool& t) = 0;
     virtual Matrix getGradients() = 0;
-    virtual unsigned size() const = 0;
-    virtual void init(unsigned nbInputs, unsigned nbOutputs, unsigned batchSize, std::mt19937& generator) = 0;
+    virtual size_t size() const = 0;
+    virtual void init(size_t nbInputs, size_t nbOutputs, size_t batchSize, std::mt19937& generator) = 0;
     virtual void updateWeights(double learningRate, double L1, double L2, Optimizer opti, double momentum, double window, ThreadPool& t) = 0;
     virtual void save() = 0;
     virtual void loadSaved() = 0;
@@ -93,13 +93,13 @@ public:
     }
 
 
-    void init(unsigned nbInputs, unsigned nbOutputs, unsigned batchSize, std::mt19937& generator)
+    void init(size_t nbInputs, size_t nbOutputs, size_t batchSize, std::mt19937& generator)
     {
         _inputSize = nbInputs;
         _batchSize = batchSize;
         try
         {
-            for(unsigned i = 0; i < size(); i++)
+            for(size_t i = 0; i < size(); i++)
             {
                 _neurons[i].init(_param.distrib, _param.mean_boundary, _param.deviation, nbInputs, nbOutputs, batchSize, _param.k, generator, _param.useOutput);
             }
@@ -117,17 +117,17 @@ public:
         Matrix output(inputs.lines(), Vector(_neurons.size(), 0));
         std::vector<std::future<void>> tasks;
 
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             tasks.push_back(t.enqueue([this, &inputs, &output, i]()->void
             {
                 //one result per feature (for each neuron)
                 Vector result = _neurons[i].process(inputs);
-                for(unsigned j = 0; j < result.size(); j++)
+                for(size_t j = 0; j < result.size(); j++)
                     output[j][i] = result[j];
             }));
         }
-        for(unsigned i = 0; i < tasks.size(); i++)
+        for(size_t i = 0; i < tasks.size(); i++)
         {
             tasks[i].get();
         }
@@ -141,13 +141,13 @@ public:
         Matrix output(_batchSize, Vector(_neurons.size(), 0));
         std::vector<std::future<void>> tasks;
 
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             tasks.push_back(t.enqueue([this, &inputs, &output, i, dropout, dropconnect, &dropoutDist, &dropconnectDist, &dropGen]()->void
             {
                 //one result per feature (for each neuron)
                 Vector result = _neurons[i].processToLearn(inputs, dropconnect, dropconnectDist, dropGen);
-                for(unsigned j = 0; j < result.size(); j++)
+                for(size_t j = 0; j < result.size(); j++)
                 {
                     output[j][i] = result[j];
                     //dropOut
@@ -161,7 +161,7 @@ public:
                 }
             }));
         }
-        for(unsigned i = 0; i < tasks.size(); i++)
+        for(size_t i = 0; i < tasks.size(); i++)
         {
             tasks[i].get();
         }
@@ -173,14 +173,14 @@ public:
     {
         std::vector<std::future<void>> tasks;
 
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             tasks.push_back(t.enqueue([this, &inputGradients, i]()->void
             {
                 _neurons[i].computeGradients(inputGradients[i]);
             }));
         }
-        for(unsigned i = 0; i < tasks.size(); i++)
+        for(size_t i = 0; i < tasks.size(); i++)
         {
             tasks[i].get();
         }
@@ -189,7 +189,7 @@ public:
 
     void save()
     {
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             _neurons[i].save();
         }
@@ -198,7 +198,7 @@ public:
 
     void loadSaved()
     {
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             _neurons[i].loadSaved();
         }
@@ -210,12 +210,12 @@ public:
     {
         Matrix grad(_inputSize, Vector(_batchSize, 0));
 
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             Matrix neuronGrad = _neurons[i].getGradients();
-            for(unsigned j = 0; j < neuronGrad.lines(); j++)
+            for(size_t j = 0; j < neuronGrad.lines(); j++)
             {
-                for(unsigned k = 0; k < neuronGrad.columns(); k++)
+                for(size_t k = 0; k < neuronGrad.columns(); k++)
                     grad[k][j] += neuronGrad[j][k];
             }
         }
@@ -228,21 +228,21 @@ public:
     {
         std::vector<std::future<void>> tasks;
 
-        for(unsigned i = 0; i < _neurons.size(); i++)
+        for(size_t i = 0; i < _neurons.size(); i++)
         {
             tasks.push_back(t.enqueue([=]()->void
             {
                 _neurons[i].updateWeights(learningRate, L1, L2, _param.maxNorm, opti, momentum, window);
             }));
         }
-        for(unsigned i = 0; i < tasks.size(); i++)
+        for(size_t i = 0; i < tasks.size(); i++)
         {
             tasks[i].get();
         }
     }
 
 
-    unsigned size() const
+    size_t size() const
     {
         return _neurons.size();
     }
@@ -253,14 +253,14 @@ public:
         std::vector<std::pair<Matrix, Vector>> weights(size());
         std::vector<std::future<void>> tasks;
 
-        for(unsigned i = 0; i < size(); i++)
+        for(size_t i = 0; i < size(); i++)
         {
             tasks.push_back(t.enqueue([this, &weights, i]()->void
             {
                 weights[i] = _neurons[i].getWeights();
             }));
         }
-        for(unsigned i = 0; i < tasks.size(); i++)
+        for(size_t i = 0; i < tasks.size(); i++)
         {
             tasks[i].get();
         }
@@ -271,8 +271,8 @@ public:
 protected:
 
     LayerParam _param;
-    unsigned _inputSize;
-    unsigned _batchSize;
+    size_t _inputSize;
+    size_t _batchSize;
     std::vector<Neuron<Aggr_t, Act_t>> _neurons;
 };
